requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.7
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
coverage==5.5
Django==3.1.3
django-cors-headers==3.10.0
django-filter==2.4.0
djangorestframework==3.12.2
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
protobuf==3.17.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyparsing==2.4.7
python-dotenv==0.18.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.0.1
wrapt==1.12.1
stripe>=2.48.0,<3.0
